# docs/openapi/agent-service.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Investigation Agent Service",
    "description": "Synchronous bounded investigations. POST waits for terminal status. GET retrieves process-local state. No background queue; restart loses results. Health is liveness only.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agent/investigations": {
      "post": {
        "summary": "Create",
        "operationId": "create_api_v1_agent_investigations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Invalid question or UUID; sanitized validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "413": {
            "description": "Request exceeds 16 KiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Sanitized internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "503": {
            "description": "Investigation capacity exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "description": "Valid UUIDs are propagated; invalid/missing values are replaced.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "Waits up to the configured workflow deadline (60 seconds by default). Returns 201 for a created completed, partial or failed investigation; inspect status and error."
      }
    },
    "/api/v1/agent/investigations/{identity}": {
      "get": {
        "summary": "Get",
        "operationId": "get_api_v1_agent_investigations__identity__get",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Identity"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "description": "Valid UUIDs are propagated; invalid/missing values are replaced.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Invalid question or UUID; sanitized validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "413": {
            "description": "Request exceeds 16 KiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Sanitized internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Investigation missing, expired or lost after restart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/internal/v1/investigations": {
      "post": {
        "summary": "Create",
        "operationId": "internal_create_api_v1_agent_investigations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Invalid question or UUID; sanitized validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "413": {
            "description": "Request exceeds 16 KiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Sanitized internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "503": {
            "description": "Investigation capacity exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "description": "Valid UUIDs are propagated; invalid/missing values are replaced.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "Waits up to the configured workflow deadline (60 seconds by default). Returns 201 for a created completed, partial or failed investigation; inspect status and error.",
        "tags": [
          "Internal Java integration"
        ],
        "x-internal": true
      }
    },
    "/internal/v1/investigations/{identity}": {
      "get": {
        "summary": "Get",
        "operationId": "internal_get_api_v1_agent_investigations__identity__get",
        "parameters": [
          {
            "name": "identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Identity"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "description": "Valid UUIDs are propagated; invalid/missing values are replaced.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "422": {
            "description": "Invalid question or UUID; sanitized validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "413": {
            "description": "Request exceeds 16 KiB",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "500": {
            "description": "Sanitized internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "Investigation missing, expired or lost after restart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "description": "Propagated or generated request UUID.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        },
        "tags": [
          "Internal Java integration"
        ],
        "x-internal": true
      }
    }
  },
  "components": {
    "schemas": {
      "Evidence": {
        "properties": {
          "type": {
            "type": "string",
            "const": "sql_result",
            "title": "Type",
            "default": "sql_result"
          },
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "rowCount": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Rowcount",
            "default": 0
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "sqlQueryId": {
            "type": "string",
            "title": "Sqlqueryid"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "rows": {
            "items": {
              "items": {},
              "type": "array"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "description",
          "sqlQueryId",
          "columns",
          "rows"
        ],
        "title": "Evidence"
      },
      "Finding": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title",
            "default": "Observation"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.0
          },
          "description": {
            "type": "string",
            "maxLength": 8000,
            "title": "Description"
          },
          "evidenceIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Evidenceids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "description",
          "evidenceIds"
        ],
        "title": "Finding"
      },
...
      "Request": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "Request"
      },
      "Response": {
        "properties": {
          "investigationId": {
            "type": "string",
            "format": "uuid",
            "title": "Investigationid"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "partial",
              "failed"
            ],
            "title": "Status"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "sqlQueries": {
            "items": {
              "$ref": "#/components/schemas/SqlQuery"
            },
            "type": "array",
            "title": "Sqlqueries"
          },
          "toolExecutions": {
            "items": {
              "$ref": "#/components/schemas/ToolExecution"
            },
            "type": "array",
            "title": "Toolexecutions"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SafeError"
              },
              {
                "type": "null"
              }
            ]
          },
          "currentStep": {
            "type": "string",
            "title": "Currentstep"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Steps"
          },
          "probableRootCause": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probablerootcause"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "dataSource": {
            "type": "string",
            "title": "Datasource"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "investigationId",
          "status",
          "question",
          "summary",
          "findings",
          "evidence",
          "sqlQueries",
          "toolExecutions",
          "createdAt",
          "completedAt",
          "error",
          "currentStep",
          "steps",
          "probableRootCause",
          "confidence",
          "limitations",
          "mode",
          "dataSource"
        ],
        "title": "Response"
      },
      "SafeError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "SafeError"
      },
      "SqlQuery": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sql": {
            "type": "string",
            "title": "Sql"
          },
          "rowCount": {
            "type": "integer",
            "title": "Rowcount"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "sql",
          "rowCount",
          "truncated"
        ],
        "title": "SqlQuery"
      },
      "ToolExecution": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "toolName": {
            "type": "string",
            "title": "Toolname"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Startedat"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Completedat"
          },
          "durationMs": {
            "type": "integer",
            "title": "Durationms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "toolName",
          "status",
          "startedAt",
          "completedAt",
          "durationMs"
        ],
        "title": "ToolExecution"
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error",
          "correlationId"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  }
}